UserEntity -> Optional<UserEntity>, service refactored
