+++ src/main/java/com/example/rightCity/repository/UserRepository.java
@@ -3,6 +3,8 @@ package com.example.rightCity.repository;
 import com.example.rightCity.entity.UserEntity;
 import org.springframework.data.repository.CrudRepository;
 
+import java.util.Optional;
+
 public interface UserRepository extends CrudRepository<UserEntity, Long> {
-    UserEntity findByMail(String mail);
+    Optional<UserEntity> findByMail(String mail);
 }
