+++ src/test/java/com/example/rightCity/repository/UserRepositoryTest.java
@@ -5,13 +5,11 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
-import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.context.junit.jupiter.SpringExtension;
 
 import javax.transaction.Transactional;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.junit.jupiter.api.Assertions.*;
 
 @DataJpaTest
 @Transactional
@@ -34,7 +32,7 @@ class UserRepositoryTest {
         expected.setPassword("0uCOQvqKPASSWORD");
 
         //when
-        UserEntity actual = userRepository.findByMail(email);
+        UserEntity actual = userRepository.findByMail(email).orElseThrow();
 
         //then
         assertThat(actual).isEqualTo(expected);
